+++ lib/vagrant/devcommands/messages.rb
@@ -10,6 +10,11 @@ module VagrantPlugins
         out.call I18n.t("#{I18N_KEY}.chain_no_help")
       end
 
+      def self.deprecated_box_config(&out)
+        out.call I18n.t("#{I18N_KEY}.deprecated_box_config")
+        out.call ''
+      end
+
       def self.command_no_help(&out)
         out.call I18n.t("#{I18N_KEY}.command_no_help")
       end