+++ lib/vagrant/devcommands/model/command.rb
@@ -16,6 +16,8 @@ module VagrantPlugins
         attr_reader :help
         attr_reader :usage
 
+        attr_reader :deprecated_box_config
+
         def initialize(spec)
           @name = spec[:name]
 
@@ -27,6 +29,8 @@ module VagrantPlugins
           @desc    = spec[:desc]
           @help    = spec[:help]
           @usage   = spec[:usage]
+
+          @deprecated_box_config = spec.key?(:box)
         end
 
         def run_script(argv)