+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.9.0.dev
+
+- Deprecations
+  - The configuration parameter `:box` has been renamed to `:machine` to
+    match the vagrant naming. Support for the old configuration will be
+    dropped in a future version
+
 ## v0.8.0 (2017-04-19)
 
 - Enhancements